Revert " Framework: Async load all Web Previews (#20960)"
This reverts commit a839c1ba72cf6749647f2fe3e917eb906193cd7a.
